Move 'rand_selected_interface' to dual_tor_utils
