+++ tests/dualtor/conftest.py
@@ -34,13 +34,3 @@ def verify_crm_nexthop_counter_not_increased(lower_tor_host):
     yield
     diff = get_crm_nexthop_counter(lower_tor_host) - original_counter
     py_assert(diff == 0, "crm nexthop counter is increased by {}.".format(diff))
-
-
-@pytest.fixture(scope="function")
-def rand_selected_interface(rand_selected_dut):
-    """Select a random interface to test."""
-    tor = rand_selected_dut
-    server_ips = mux_cable_server_ip(tor)
-    iface = str(random.choice(server_ips.keys()))
-    logging.info("select DUT interface %s to test.", iface)
-    return iface, server_ips[iface]
